experiment with vm stuff

+++ files/.config/zsh/custom/101-function.zsh
@@ -44,28 +44,32 @@ __nmap_iface() {
 }
 
 # start looking-glass
-game() {
-    echo "Stopping services"
-    sudo systemctl stop cronie
-    systemctl --user stop redshift compton wallpaper
+play-vm() {
+    if ! [ -S /tmp/win10.sock ]; then
+        echo "Cannot find spice socket! Is the vm started?"
+    else
+        echo "Stopping services"
+        sudo systemctl stop cronie
+        systemctl --user stop redshift compton wallpaper
 
-    echo "Setup cpuset cgroup for host"
-    sudo cset set -c 0,4 -s system
-    sudo cset proc -m -f root -t system
-    sudo cset proc -k -f root -t system --force
+        echo "Setup cpuset cgroup for host"
+        sudo cset set -c 0,4 -s system
+        sudo cset proc -m -f root -t system
+        sudo cset proc -k -f root -t system --force
 
-    echo "Setup cpumask"
-    sudo bash -c "echo 5 > /sys/devices/virtual/workqueue/cpumask"
+        echo "Setup cpumask"
+        sudo bash -c "echo 5 > /sys/devices/virtual/workqueue/cpumask"
 
-    echo "Starting looking-glass"
-    LD_PRELOAD=/usr/\$LIB/libgamemodeauto.so looking-glass-client -o opengl:preventBuffer=0 -MFk
+        echo "Starting looking-glass"
+        LD_PRELOAD=/usr/\$LIB/libgamemodeauto.so looking-glass-client -p 0 -c /tmp/win10.sock -o opengl:preventBuffer=0 -MF
 
-    echo "Restore system"
-    # cpumask
-    sudo bash -c "echo f > /sys/devices/virtual/workqueue/cpumask"
-    # cpuset
-    sudo cset set -d system &>/dev/null
-    # services
-    systemctl --user start redshift compton wallpaper
-    sudo systemctl start cronie
+        echo "Restore system"
+        # cpumask
+        sudo bash -c "echo f > /sys/devices/virtual/workqueue/cpumask"
+        # cpuset
+        sudo cset set -d system &>/dev/null
+        # services
+        systemctl --user start redshift compton wallpaper
+        sudo systemctl start cronie
+    fi
 }
